refactor: create profile flow navigation, rework mining calculator to separate concepts
