+++ src/components/Slider/index.tsx
