+++ src/screens/ProfileFlow/MyBadges/components/CheckMark.tsx
