+++ src/screens/ProfileFlow/MyBadges/index.tsx
@@ -1,0 +1,39 @@
+// SPDX-License-Identifier: BUSL-1.1
+
+import {COLORS} from '@constants/colors';
+import {useScrollShadow} from '@hooks/useScrollShadow';
+import {Header} from '@navigation/components/Header';
+import {FaqButton} from '@navigation/components/Header/components/FaqButton';
+import {useFocusStatusBar} from '@navigation/hooks/useFocusStatusBar';
+import {t} from '@utils/i18n';
+import React from 'react';
+import {StyleSheet, View} from 'react-native';
+import Animated from 'react-native-reanimated';
+
+export const MyBadges = () => {
+  useFocusStatusBar({style: 'dark-content'});
+  const {scrollHandler, shadowStyle} = useScrollShadow();
+
+  return (
+    <View style={styles.container}>
+      <Header
+        containerStyle={shadowStyle}
+        color={COLORS.darkBlue}
+        backgroundColor={COLORS.white}
+        renderRightButtons={FaqButton}
+        title={t('my_badges.title')}
+      />
+      <Animated.ScrollView
+        onScroll={scrollHandler}
+        scrollEventThrottle={16}
+        showsVerticalScrollIndicator={false}
+      />
+    </View>
+  );
+};
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+  },
+});
