+++ src/screens/ProfileFlow/MyRoles/components/Role.tsx
@@ -1,0 +1,80 @@
+// SPDX-License-Identifier: BUSL-1.1
+
+import {COLORS} from '@constants/colors';
+import {FONTS} from '@constants/fonts';
+import {SCREEN_SIDE_OFFSET} from '@constants/styles';
+import {CheckMark} from '@screens/ProfileFlow/MyRoles/components/CheckMark';
+import React, {ReactNode} from 'react';
+import {StyleSheet, Text, View} from 'react-native';
+import {SvgProps} from 'react-native-svg';
+import {font, rem} from 'rn-units';
+
+type Props = {
+  title: string;
+  tagline: string;
+  description: string;
+  renderIcon: (props: SvgProps) => ReactNode;
+  backgroundColor?: string;
+  checked?: boolean;
+};
+
+export const Role = ({
+  title,
+  tagline,
+  description,
+  renderIcon,
+  backgroundColor,
+  checked = false,
+}: Props) => {
+  return (
+    <View style={[styles.container, {backgroundColor}]}>
+      <Text style={styles.titleText}>{title}</Text>
+      <View>
+        {renderIcon({style: styles.icon})}
+        {checked && <CheckMark style={styles.checkmark} />}
+      </View>
+      <Text style={styles.taglineText}>{tagline}</Text>
+      <Text style={styles.descriptionText}>{description}</Text>
+    </View>
+  );
+};
+
+const styles = StyleSheet.create({
+  container: {
+    alignItems: 'center',
+    paddingBottom: rem(22),
+    paddingHorizontal: SCREEN_SIDE_OFFSET + rem(30),
+  },
+  titleText: {
+    marginTop: rem(20),
+    fontFamily: FONTS.primary.bold,
+    fontSize: font(20),
+    color: COLORS.darkBlue,
+  },
+  icon: {
+    marginTop: rem(12),
+    width: rem(150),
+    height: rem(140),
+  },
+  checkmark: {
+    position: 'absolute',
+    bottom: rem(5),
+    right: rem(10),
+  },
+  taglineText: {
+    marginTop: rem(8),
+    fontFamily: FONTS.primary.bold,
+    fontSize: font(18),
+    lineHeight: font(20),
+    color: COLORS.darkBlue,
+    textAlign: 'center',
+  },
+  descriptionText: {
+    marginTop: rem(14),
+    fontFamily: FONTS.primary.regular,
+    fontSize: font(14),
+    lineHeight: font(20),
+    color: COLORS.greyText,
+    textAlign: 'center',
+  },
+});
